# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tajweed Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: Courses list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a course by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
        "404":
          description: Not found
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Lessons ─────────────────────────────────────────────────────────────────
  /lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List all lessons (always accessible to authenticated users)
      parameters:
        - name: level
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Lessons list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create lesson (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{id}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get lesson by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
        "404":
          description: Not found
    patch:
      operationId: updateLesson
      tags: [lessons]
      summary: Update lesson (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LessonUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete lesson (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Tajweed Rules ────────────────────────────────────────────────────────────
  /tajweed-rules:
    get:
      operationId: listTajweedRules
      tags: [tajweedRules]
      summary: List all tajweed rules (always accessible)
      responses:
        "200":
          description: Rules list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TajweedRule"
    post:
      operationId: createTajweedRule
      tags: [tajweedRules]
      summary: Create tajweed rule (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TajweedRuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TajweedRule"
  /tajweed-rules/{id}:
    patch:
      operationId: updateTajweedRule
      tags: [tajweedRules]
      summary: Update tajweed rule (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TajweedRuleInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TajweedRule"
    delete:
      operationId: deleteTajweedRule
      tags: [tajweedRules]
      summary: Delete tajweed rule (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Students (admin management) ──────────────────────────────────────────────
...
  /enrollments:
    get:
      operationId: listEnrollments
      tags: [enrollments]
      summary: List enrollments (admin sees all, student sees own)
      parameters:
        - name: courseId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, active, completed, cancelled]
      responses:
        "200":
          description: Enrollments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: createEnrollment
      tags: [enrollments]
      summary: Create enrollment (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}:
    get:
      operationId: getEnrollment
      tags: [enrollments]
      summary: Get enrollment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrollment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
    patch:
      operationId: updateEnrollment
      tags: [enrollments]
      summary: Update enrollment status (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"

  # ── Assignments ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [teacher, student]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, clerkId, email, name, role, createdAt]
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        level:
          type: integer
          description: "1 = Level 1, 2 = Level 2"
        durationDays:
          type: integer
        price:
          type: number
        capacity:
          type: integer
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        isActive:
          type: boolean
        enrolledCount:
          type: integer
        createdAt:
          type: string
      required: [id, title, description, level, durationDays, price, capacity, isActive, enrolledCount, createdAt]
    CourseInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        level:
          type: integer
        durationDays:
          type: integer
        price:
          type: number
        capacity:
          type: integer
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
      required: [title, description, level, durationDays, price, capacity]
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        durationDays:
          type: integer
        price:
          type: number
        capacity:
          type: integer
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        isActive:
          type: boolean

    # ── Lessons ──────────────────────────────────────────────────────────────
    Lesson:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        level:
          type: integer
        orderIndex:
          type: integer
        ruleId:
          type: ["integer", "null"]
        videoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, content, level, orderIndex, createdAt]
    LessonInput:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        level:
          type: integer
        orderIndex:
          type: integer
        ruleId:
          type: ["integer", "null"]
        videoUrl:
          type: ["string", "null"]
      required: [title, content, level, orderIndex]
    LessonUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        orderIndex:
          type: integer
        ruleId:
          type: ["integer", "null"]
        videoUrl:
          type: ["string", "null"]

    # ── Tajweed Rules ─────────────────────────────────────────────────────────
    TajweedRule:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        description:
          type: string
        example:
          type: string
        level:
          type: integer
        orderIndex:
          type: integer
        createdAt:
          type: string
      required: [id, nameAr, nameEn, description, example, level, orderIndex, createdAt]
    TajweedRuleInput:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        description:
          type: string
        example:
          type: string
        level:
          type: integer
        orderIndex:
          type: integer
      required: [nameAr, nameEn, description, example, level, orderIndex]

    # ── Enrollments ──────────────────────────────────────────────────────────
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        student:
          $ref: "#/components/schemas/User"
        course:
          $ref: "#/components/schemas/Course"
        createdAt:
          type: string
      required: [id, studentId, courseId, status, createdAt]
    EnrollmentInput:
      type: object
      properties:
        studentId:
          type: integer
        courseId:
          type: integer
        startDate:
          type: ["string", "null"]
      required: [studentId, courseId]
    EnrollmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]

    # ── Assignments ──────────────────────────────────────────────────────────
...